+++ tests/fixtures/index.js
@@ -1,104 +1,20 @@
 import realProv from './real-prov.json';
 import realCwl from './real-cwl.json';
+import complexProv from './complex-prov';
 import complexCwl from './complex-cwl.json';
+import simpleProv from './simple-prov';
 import simpleCwl from './simple-cwl.json';
 
 export const simple = {
   getNameForActivity: (id) => id,
   getNameForEntity: (id) => id,
-  prov: {
-    prefix: {
-      ex: 'http://example.com',
-    },
-    entity: {
-      'ex:input': { 'prov:label': 'Input', 'ex:note': 'Begins here...' },
-      'ex:output': { 'prov:label': 'Output', 'ex:note': '... and ends here.' },
-    },
-    activity: {
-      'ex:process': { 'prov:label': 'Process' },
-    },
-    wasGeneratedBy: {
-      '_:1': {
-        'prov:activity': 'ex:process',
-        'prov:entity': 'ex:output',
-      },
-    },
-    used: {
-      '_:2': {
-        'prov:activity': 'ex:process',
-        'prov:entity': 'ex:input',
-      },
-    },
-  },
+  prov: simpleProv,
   cwl: simpleCwl,
 };
 
 export const complex = {
   getNameForActivity: (id) => id,
-  prov: {
-    /*
-    E1 ═> A1 ═> E3 ═> A3 ═> E5
-    ║╚═══════╗  ╚══════╗
-    ╚═════╗  ╚════════╗║
-          V           VV
-    E2 ═> A2 ═> E4 ═> A4 ═> E6
-          ║
-          ╚═══> E7
-    */
-
-    /* eslint-disable object-curly-newline */
-    /* eslint-disable indent */
-    prefix: {
-      hubmap: 'https://hubmapconsortium.org',
-    },
-    entity: {
-      'hubmap:ent-1': { 'prov:label': 'ent-1' },
-      'hubmap:ent-2': { 'prov:label': 'ent-2' },
-      'hubmap:ent-3': { 'prov:label': 'ent-3' },
-      'hubmap:ent-4': { 'prov:label': 'ent-4' },
-      'hubmap:ent-5': { 'prov:label': 'ent-5' },
-      'hubmap:ent-6': { 'prov:label': 'ent-6' },
-      'hubmap:ent-7': { 'prov:label': 'ent-7' },
-    },
-    activity: {
-      'hubmap:act-1': { 'prov:label': 'act-1' },
-      'hubmap:act-2': { 'prov:label': 'act-2' },
-      'hubmap:act-3': { 'prov:label': 'act-3' },
-      'hubmap:act-4': { 'prov:label': 'act-4' },
-    },
-    wasGeneratedBy: {
-      '_:1': { 'prov:entity': 'hubmap:ent-3',
-               'prov:activity': 'hubmap:act-1' },
-      '_:2': { 'prov:entity': 'hubmap:ent-4',
-               'prov:activity': 'hubmap:act-2' },
-      '_:3': { 'prov:entity': 'hubmap:ent-5',
-               'prov:activity': 'hubmap:act-3' },
-      '_:4': { 'prov:entity': 'hubmap:ent-6',
-               'prov:activity': 'hubmap:act-4' },
-      '_:5': { 'prov:entity': 'hubmap:ent-7',
-               'prov:activity': 'hubmap:act-2' },
-    },
-    used: {
-      '_:5': { 'prov:activity': 'hubmap:act-1',
-               'prov:entity': 'hubmap:ent-1' },
-
-      '_:6': { 'prov:activity': 'hubmap:act-2',
-               'prov:entity': 'hubmap:ent-1' },
-      '_:7': { 'prov:activity': 'hubmap:act-2',
-               'prov:entity': 'hubmap:ent-2' },
-
-      '_:8': { 'prov:activity': 'hubmap:act-3',
-               'prov:entity': 'hubmap:ent-3' },
-
-      '_:9': { 'prov:activity': 'hubmap:act-4',
-               'prov:entity': 'hubmap:ent-1' },
-      '_:10': { 'prov:activity': 'hubmap:act-4',
-                'prov:entity': 'hubmap:ent-3' },
-      '_:11': { 'prov:activity': 'hubmap:act-4',
-                'prov:entity': 'hubmap:ent-4' },
-      },
-    },
-    /* eslint-enable */
+  prov: complexProv,
   cwl: complexCwl,
 };
 
@@ -117,6 +33,5 @@ export const real = {
   cwl: realCwl,
 };
 
-// The demos reference the default export:
-// Update this to update all the demos.
+// The React demo references the default export.
 export default complex;