+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## 0.0.3 - in progress
+### Changed
+- Fixed redundant fixtures.
 
 ## [0.0.2](https://www.npmjs.com/package/@hubmap/prov-vis/v/0.0.2) - 2019-11-04
 ### Added